Adds Seek and Prev to Iterator interface (#11)
Although primarily a leveldb iterator interface, having Prev in addition
to Next is going to be a helpful utility if we can implement it with
other engines. The larger requirement is Seek which is absolutely needed
for pagination and other advanced iteration. These are primarily a
passthrough for leveldb operations, but I've also added some tests for
the empty iterator which is not a passthrough.