Allow Engine-Specific Options in Open
This PR modifies the Honu config to allow engine-specific options to be
passed in at Open, and in particular to allow leveldb options to be
passed in on open. The new configuration mechanism is similar to the
access Options mechanism, a Honu configuration can be created without
any options and a default configuration is returned. Using WithLevelDB
or WithReplica, other options can be specified to the config.

This PR also removes some technical debt with the Honu replica
configuration that was ported over from a previous project.