Update Version Checks (#18)
This PR implements a required change for Trtl Anti-Entropy: Update
checks if the version being updated is still later than the local
version. I've also added a WithForce() option that bypasses checks if
the user wants to force put an object to the database, overriding the
namespace and version checks.