Skip, stomp, and linear version history tracking in Update (#19)
Adds an UpdateType return value to the db.Update method so callers can
track what happens to the underlying version history. The types are
directly connected to new Version comparison methods of similar names,
which will hopefully make it easier for us to reason about whats
happening in the version history and in anti-entropy in general.