Better Tombstone handling in Iter (#21)
Adds an option for iterating with tombstones but by default the Iter
function skips tombstones to match the Get functionality. Also fixes the
colons in namespace issue by removing the namespace prefix rather than
splitting on the separator. Increases tests to ensure that Iter covers
undead objects.

Note, this will require a change to Trtl replication, it must use the
WithTombstone option when it is iterating in initiator phase 1 and
remote phase 2; otherwise deletes will not be replicated.